move tests and fix name
