+++ core/cmd/local_client_test.go
@@ -156,7 +156,7 @@ KEEPER_REGISTRY_SYNC_INTERVAL:
 KEEPER_REGISTRY_SYNC_UPKEEP_QUEUE_SIZE: 0
 KEEPER_CHECK_UPKEEP_GAS_PRICE_FEATURE_ENABLED: false
 KEEPER_TURN_LOOK_BACK: 1000
-KEEPER_TURN_FLAG: false
+KEEPER_TURN_FLAG_ENABLED: false
 LEASE_LOCK_DURATION: 10s
 LEASE_LOCK_REFRESH_INTERVAL: 1s
 FLAGS_CONTRACT_ADDRESS:
