+++ core/services/keeper/orm_test.go
@@ -1,6 +1,8 @@
 package keeper_test
 
 import (
+	"fmt"
+	"sort"
 	"testing"
 	"time"
 
@@ -17,6 +19,7 @@ import (
 	"github.com/smartcontractkit/chainlink/core/internal/testutils/pgtest"
 	"github.com/smartcontractkit/chainlink/core/logger"
 	"github.com/smartcontractkit/chainlink/core/services/keeper"
+	"github.com/smartcontractkit/chainlink/core/utils"
 )
 
 var (
@@ -167,8 +170,20 @@ func TestKeeperDB_EligibleUpkeeps_Shuffle(t *testing.T) {
 	}
 	assert.NotEqualValues(t, ordered, shuffled)
 }
+func TestKeeperDB_RegistryByContractAddress(t *testing.T) {
+	t.Parallel()
+	db, config, orm := setupKeeperDB(t)
+	ethKeyStore := cltest.NewKeyStore(t, db, config).Eth()
 
-func TestKeeperDB_EligibleUpkeeps_BlockCountPerTurn(t *testing.T) {
+	registry, _ := cltest.MustInsertKeeperRegistry(t, db, orm, ethKeyStore, 0, 1, 20)
+	cltest.MustInsertKeeperRegistry(t, db, orm, ethKeyStore, 0, 1, 20)
+
+	registryByContractAddress, err := orm.RegistryByContractAddress(registry.ContractAddress)
+	require.NoError(t, err)
+	require.Equal(t, registry, registryByContractAddress)
+}
+
+func TestKeeperDB_NewEligibleUpkeeps_Shuffle(t *testing.T) {
 	t.Parallel()
 	db, config, orm := setupKeeperDB(t)
 	ethKeyStore := cltest.NewKeyStore(t, db, config).Eth()
@@ -178,117 +193,59 @@ func TestKeeperDB_EligibleUpkeeps_BlockCountPerTurn(t *testing.T) {
 
 	registry, _ := cltest.MustInsertKeeperRegistry(t, db, orm, ethKeyStore, 0, 1, 20)
 
-	upkeeps := [5]keeper.UpkeepRegistration{
-		newUpkeep(registry, 0),
-		newUpkeep(registry, 1),
-		newUpkeep(registry, 2),
-		newUpkeep(registry, 3),
-		newUpkeep(registry, 4),
-	}
-
-	upkeeps[0].LastRunBlockHeight = 0  // Never run
-	upkeeps[1].LastRunBlockHeight = 41 // Run last turn, outside grade period
-	upkeeps[2].LastRunBlockHeight = 46 // Run last turn, outside grade period
-	upkeeps[3].LastRunBlockHeight = 59 // Run last turn, inside grace period (EXCLUDE)
-	upkeeps[4].LastRunBlockHeight = 61 // Run this turn, inside grace period (EXCLUDE)
-
-	for i, upkeep := range upkeeps {
-		upkeep.PositioningConstant = int32(i)
-		err := orm.UpsertUpkeep(&upkeep)
+	ordered := [100]int64{}
+	for i := 0; i < 100; i++ {
+		k := newUpkeep(registry, int64(i))
+		ordered[i] = int64(i)
+		err := orm.UpsertUpkeep(&k)
 		require.NoError(t, err)
 	}
+	cltest.AssertCount(t, db, "upkeep_registrations", 100)
 
-	cltest.AssertCount(t, db, "upkeep_registrations", 5)
-
-	eligibleUpkeeps, err := orm.EligibleUpkeepsForRegistry(registry.ContractAddress, blockheight, gracePeriod)
+	eligibleUpkeeps, err := orm.NewEligibleUpkeepsForRegistry(registry.ContractAddress, blockheight, gracePeriod, fmt.Sprintf("%b", utils.NewHash().Big()))
 	assert.NoError(t, err)
 
-	// 3 out of 5 are eligible, check that ids are 0,1 or 2 but order is shuffled so can not use equals
-	require.Len(t, eligibleUpkeeps, 3)
-	assert.Less(t, eligibleUpkeeps[0].UpkeepID, int64(3))
-	assert.Less(t, eligibleUpkeeps[1].UpkeepID, int64(3))
-	assert.Less(t, eligibleUpkeeps[2].UpkeepID, int64(3))
-
-	// preloads registry data
-	assert.Equal(t, registry.ID, eligibleUpkeeps[0].RegistryID)
-	assert.Equal(t, registry.ID, eligibleUpkeeps[1].RegistryID)
-	assert.Equal(t, registry.ID, eligibleUpkeeps[2].RegistryID)
-	assert.Equal(t, registry.CheckGas, eligibleUpkeeps[0].Registry.CheckGas)
-	assert.Equal(t, registry.CheckGas, eligibleUpkeeps[1].Registry.CheckGas)
-	assert.Equal(t, registry.CheckGas, eligibleUpkeeps[2].Registry.CheckGas)
-	assert.Equal(t, registry.ContractAddress, eligibleUpkeeps[0].Registry.ContractAddress)
-	assert.Equal(t, registry.ContractAddress, eligibleUpkeeps[1].Registry.ContractAddress)
-	assert.Equal(t, registry.ContractAddress, eligibleUpkeeps[2].Registry.ContractAddress)
-}
-
-func TestKeeperDB_EligibleUpkeeps_GracePeriod(t *testing.T) {
-	t.Parallel()
-	db, config, orm := setupKeeperDB(t)
-	ethKeyStore := cltest.NewKeyStore(t, db, config).Eth()
-
-	blockheight := int64(120)
-	gracePeriod := int64(100)
-
-	registry, _ := cltest.MustInsertKeeperRegistry(t, db, orm, ethKeyStore, 0, 2, 20)
-	upkeep1 := newUpkeep(registry, 0)
-	upkeep1.LastRunBlockHeight = 0
-	upkeep2 := newUpkeep(registry, 1)
-	upkeep2.LastRunBlockHeight = 19
-	upkeep3 := newUpkeep(registry, 2)
-	upkeep3.LastRunBlockHeight = 20
-
-	for _, upkeep := range [3]keeper.UpkeepRegistration{upkeep1, upkeep2, upkeep3} {
-		err := orm.UpsertUpkeep(&upkeep)
-		require.NoError(t, err)
+	require.Len(t, eligibleUpkeeps, 100)
+	shuffled := [100]int64{}
+	for i := 0; i < 100; i++ {
+		shuffled[i] = eligibleUpkeeps[i].UpkeepID
 	}
-
-	cltest.AssertCount(t, db, "upkeep_registrations", 3)
-
-	eligibleUpkeeps, err := orm.EligibleUpkeepsForRegistry(registry.ContractAddress, blockheight, gracePeriod)
-	assert.NoError(t, err)
-	// 2 out of 3 are eligible, check that ids are 0 or 1 but order is shuffled so can not use equals
-	assert.Len(t, eligibleUpkeeps, 2)
-	assert.Less(t, eligibleUpkeeps[0].UpkeepID, int64(2))
-	assert.Less(t, eligibleUpkeeps[1].UpkeepID, int64(2))
+	assert.NotEqualValues(t, ordered, shuffled)
 }
 
-func TestKeeperDB_EligibleUpkeeps_KeepersRotate(t *testing.T) {
+func TestKeeperDB_NewEligibleUpkeeps_GracePeriod(t *testing.T) {
 	t.Parallel()
 	db, config, orm := setupKeeperDB(t)
 	ethKeyStore := cltest.NewKeyStore(t, db, config).Eth()
 
 	registry, _ := cltest.MustInsertKeeperRegistry(t, db, orm, ethKeyStore, 0, 2, 20)
-	registry.NumKeepers = 5
-	require.NoError(t, db.Get(&registry, `UPDATE keeper_registries SET num_keepers = 5 WHERE id = $1 RETURNING *`, registry.ID))
-	cltest.MustInsertUpkeepForRegistry(t, db, config, registry)
+
+	for i := 0; i < 100; i++ {
+		cltest.MustInsertUpkeepForRegistry(t, db, config, registry)
+	}
 
 	cltest.AssertCount(t, db, "keeper_registries", 1)
-	cltest.AssertCount(t, db, "upkeep_registrations", 1)
+	cltest.AssertCount(t, db, "upkeep_registrations", 100)
 
-	// out of 5 valid block ranges, with 5 keepers, we are eligible
-	// to submit on exactly 1 of them
-	list1, err := orm.EligibleUpkeepsForRegistry(registry.ContractAddress, 20, 0)
+	// if current keeper index = 0 and all upkeeps last perform was done by index = 0 and still within grace period
+	upkeep := keeper.UpkeepRegistration{}
+	require.NoError(t, db.Get(&upkeep, `UPDATE upkeep_registrations SET last_keeper_index = 0, last_run_block_height = 10 RETURNING *`))
+	list0, err := orm.NewEligibleUpkeepsForRegistry(registry.ContractAddress, 21, 100, fmt.Sprintf("%b", utils.NewHash().Big())) // none eligible
 	require.NoError(t, err)
-	list2, err := orm.EligibleUpkeepsForRegistry(registry.ContractAddress, 41, 0)
-	require.NoError(t, err)
-	list3, err := orm.EligibleUpkeepsForRegistry(registry.ContractAddress, 62, 0)
-	require.NoError(t, err)
-	list4, err := orm.EligibleUpkeepsForRegistry(registry.ContractAddress, 83, 0)
-	require.NoError(t, err)
-	list5, err := orm.EligibleUpkeepsForRegistry(registry.ContractAddress, 104, 0)
-	require.NoError(t, err)
+	require.Equal(t, 0, len(list0), "should be 0 as all last perform was done by current node")
 
-	totalEligible := len(list1) + len(list2) + len(list3) + len(list4) + len(list5)
-	require.Equal(t, 1, totalEligible)
+	// once passed grace period
+	list1, err := orm.NewEligibleUpkeepsForRegistry(registry.ContractAddress, 121, 100, fmt.Sprintf("%b", utils.NewHash().Big())) // none eligible
+	require.NoError(t, err)
+	require.NotEqual(t, 0, len(list1), "should get some eligible upkeeps now that they are outside grace period")
 }
 
-func TestKeeperDB_EligibleUpkeeps_KeepersCycleAllUpkeeps(t *testing.T) {
+func TestKeeperDB_NewEligibleUpkeeps_TurnsRandom(t *testing.T) {
 	t.Parallel()
 	db, config, orm := setupKeeperDB(t)
 	ethKeyStore := cltest.NewKeyStore(t, db, config).Eth()
 
-	registry, _ := cltest.MustInsertKeeperRegistry(t, db, orm, ethKeyStore, 0, 2, 20)
-	require.NoError(t, db.Get(&registry, `UPDATE keeper_registries SET num_keepers = 5, keeper_index = 3 WHERE id = $1 RETURNING *`, registry.ID))
+	registry, _ := cltest.MustInsertKeeperRegistry(t, db, orm, ethKeyStore, 0, 3, 10)
 
 	for i := 0; i < 1000; i++ {
 		cltest.MustInsertUpkeepForRegistry(t, db, config, registry)
@@ -297,29 +254,109 @@ func TestKeeperDB_EligibleUpkeeps_KeepersCycleAllUpkeeps(t *testing.T) {
 	cltest.AssertCount(t, db, "keeper_registries", 1)
 	cltest.AssertCount(t, db, "upkeep_registrations", 1000)
 
-	// in a full cycle, each node should be responsible for each upkeep exactly once
-	list1, err := orm.EligibleUpkeepsForRegistry(registry.ContractAddress, 20, 0) // someone eligible
+	// 3 keepers 10 block turns should be different every turn
+	list1, err := orm.NewEligibleUpkeepsForRegistry(registry.ContractAddress, 20, 100, fmt.Sprintf("%b", utils.NewHash().Big()))
 	require.NoError(t, err)
-	list2, err := orm.EligibleUpkeepsForRegistry(registry.ContractAddress, 40, 0) // someone eligible
+	list2, err := orm.NewEligibleUpkeepsForRegistry(registry.ContractAddress, 31, 100, fmt.Sprintf("%b", utils.NewHash().Big()))
 	require.NoError(t, err)
-	list3, err := orm.EligibleUpkeepsForRegistry(registry.ContractAddress, 60, 0) // someone eligible
+	list3, err := orm.NewEligibleUpkeepsForRegistry(registry.ContractAddress, 42, 100, fmt.Sprintf("%b", utils.NewHash().Big()))
 	require.NoError(t, err)
-	list4, err := orm.EligibleUpkeepsForRegistry(registry.ContractAddress, 80, 0) // someone eligible
-	require.NoError(t, err)
-	list5, err := orm.EligibleUpkeepsForRegistry(registry.ContractAddress, 100, 0) // someone eligible
+	list4, err := orm.NewEligibleUpkeepsForRegistry(registry.ContractAddress, 53, 100, fmt.Sprintf("%b", utils.NewHash().Big()))
 	require.NoError(t, err)
 
-	totalEligible := len(list1) + len(list2) + len(list3) + len(list4) + len(list5)
-	require.Equal(t, 1000, totalEligible)
+	// sort before compare
+	sort.Slice(list1, func(i, j int) bool {
+		return list1[i].UpkeepID < list1[j].UpkeepID
+	})
+	sort.Slice(list2, func(i, j int) bool {
+		return list2[i].UpkeepID < list2[j].UpkeepID
+	})
+	sort.Slice(list3, func(i, j int) bool {
+		return list3[i].UpkeepID < list3[j].UpkeepID
+	})
+	sort.Slice(list4, func(i, j int) bool {
+		return list4[i].UpkeepID < list4[j].UpkeepID
+	})
+
+	assert.NotEqual(t, list1, list2, "list1 vs list2")
+	assert.NotEqual(t, list1, list3, "list1 vs list3")
+	assert.NotEqual(t, list1, list4, "list1 vs list4")
 }
 
-func TestKeeperDB_EligibleUpkeeps_FiltersByRegistry(t *testing.T) {
+func TestKeeperDB_NewEligibleUpkeeps_SkipIfLastPerformedByCurrentKeeper(t *testing.T) {
 	t.Parallel()
 	db, config, orm := setupKeeperDB(t)
 	ethKeyStore := cltest.NewKeyStore(t, db, config).Eth()
 
-	registry1, _ := cltest.MustInsertKeeperRegistry(t, db, orm, ethKeyStore, 0, 2, 20)
-	registry2, _ := cltest.MustInsertKeeperRegistry(t, db, orm, ethKeyStore, 0, 2, 20)
+	registry, _ := cltest.MustInsertKeeperRegistry(t, db, orm, ethKeyStore, 0, 2, 20)
+
+	for i := 0; i < 100; i++ {
+		cltest.MustInsertUpkeepForRegistry(t, db, config, registry)
+	}
+
+	cltest.AssertCount(t, db, "keeper_registries", 1)
+	cltest.AssertCount(t, db, "upkeep_registrations", 100)
+
+	// if current keeper index = 0 and all upkeeps last perform was done by index = 0 then skip as it would not pass required turn taking
+	upkeep := keeper.UpkeepRegistration{}
+	require.NoError(t, db.Get(&upkeep, `UPDATE upkeep_registrations SET last_keeper_index = 0 RETURNING *`))
+	list0, err := orm.NewEligibleUpkeepsForRegistry(registry.ContractAddress, 21, 100, fmt.Sprintf("%b", utils.NewHash().Big())) // none eligible
+	require.NoError(t, err)
+	require.Equal(t, 0, len(list0), "should be 0 as all last perform was done by current node")
+}
+
+func TestKeeperDB_NewEligibleUpkeeps_CoverBuddy(t *testing.T) {
+	t.Parallel()
+	db, config, orm := setupKeeperDB(t)
+	ethKeyStore := cltest.NewKeyStore(t, db, config).Eth()
+
+	registry, _ := cltest.MustInsertKeeperRegistry(t, db, orm, ethKeyStore, 1, 2, 20)
+
+	for i := 0; i < 100; i++ {
+		cltest.MustInsertUpkeepForRegistry(t, db, config, registry)
+	}
+
+	cltest.AssertCount(t, db, "keeper_registries", 1)
+	cltest.AssertCount(t, db, "upkeep_registrations", 100)
+
+	upkeep := keeper.UpkeepRegistration{}
+	binaryHash := fmt.Sprintf("%b", utils.NewHash().Big())
+	listBefore, err := orm.NewEligibleUpkeepsForRegistry(registry.ContractAddress, 21, 100, binaryHash) // normal
+	require.NoError(t, err)
+	require.NoError(t, db.Get(&upkeep, `UPDATE upkeep_registrations SET last_keeper_index = 0 RETURNING *`))
+	listAfter, err := orm.NewEligibleUpkeepsForRegistry(registry.ContractAddress, 21, 100, binaryHash) // covering buddy
+	require.NoError(t, err)
+	require.Greater(t, len(listAfter), len(listBefore), "after our buddy runs all the performs we should have more eligible then a normal turn")
+}
+
+func TestKeeperDB_NewEligibleUpkeeps_FirstTurn(t *testing.T) {
+	t.Parallel()
+	db, config, orm := setupKeeperDB(t)
+	ethKeyStore := cltest.NewKeyStore(t, db, config).Eth()
+
+	registry, _ := cltest.MustInsertKeeperRegistry(t, db, orm, ethKeyStore, 0, 2, 20)
+
+	for i := 0; i < 100; i++ {
+		cltest.MustInsertUpkeepForRegistry(t, db, config, registry)
+	}
+
+	cltest.AssertCount(t, db, "keeper_registries", 1)
+	cltest.AssertCount(t, db, "upkeep_registrations", 100)
+
+	binaryHash := fmt.Sprintf("%b", utils.NewHash().Big())
+	// last keeper index is null to simulate a normal first run
+	listKpr0, err := orm.NewEligibleUpkeepsForRegistry(registry.ContractAddress, 21, 100, binaryHash) // someone eligible only kpr0 turn
+	require.NoError(t, err)
+	require.NotEqual(t, 0, len(listKpr0), "kpr0 should have some eligible as a normal turn")
+}
+
+func TestKeeperDB_NewEligibleUpkeeps_FiltersByRegistry(t *testing.T) {
+	t.Parallel()
+	db, config, orm := setupKeeperDB(t)
+	ethKeyStore := cltest.NewKeyStore(t, db, config).Eth()
+
+	registry1, _ := cltest.MustInsertKeeperRegistry(t, db, orm, ethKeyStore, 0, 1, 20)
+	registry2, _ := cltest.MustInsertKeeperRegistry(t, db, orm, ethKeyStore, 0, 1, 20)
 
 	cltest.MustInsertUpkeepForRegistry(t, db, config, registry1)
 	cltest.MustInsertUpkeepForRegistry(t, db, config, registry2)
@@ -327,57 +364,31 @@ func TestKeeperDB_EligibleUpkeeps_FiltersByRegistry(t *testing.T) {
 	cltest.AssertCount(t, db, "keeper_registries", 2)
 	cltest.AssertCount(t, db, "upkeep_registrations", 2)
 
-	list1, err := orm.EligibleUpkeepsForRegistry(registry1.ContractAddress, 20, 0)
+	binaryHash := fmt.Sprintf("%b", utils.NewHash().Big())
+	list1, err := orm.NewEligibleUpkeepsForRegistry(registry1.ContractAddress, 20, 100, binaryHash)
 	require.NoError(t, err)
-	list2, err := orm.EligibleUpkeepsForRegistry(registry2.ContractAddress, 20, 0)
+	list2, err := orm.NewEligibleUpkeepsForRegistry(registry2.ContractAddress, 20, 100, binaryHash)
 	require.NoError(t, err)
 
 	assert.Equal(t, 1, len(list1))
 	assert.Equal(t, 1, len(list2))
 }
 
-func TestKeeperDB_NextUpkeepID(t *testing.T) {
+func TestKeeperDB_NewSetLastRunInfoForUpkeepOnJob(t *testing.T) {
 	t.Parallel()
 	db, config, orm := setupKeeperDB(t)
 	ethKeyStore := cltest.NewKeyStore(t, db, config).Eth()
 
-	registry, _ := cltest.MustInsertKeeperRegistry(t, db, orm, ethKeyStore, 0, 2, 20)
-
-	nextID, err := orm.LowestUnsyncedID(registry.ID)
-	require.NoError(t, err)
-	require.Equal(t, int64(0), nextID)
-
-	upkeep := newUpkeep(registry, 0)
-	err = orm.UpsertUpkeep(&upkeep)
-	require.NoError(t, err)
-
-	nextID, err = orm.LowestUnsyncedID(registry.ID)
-	require.NoError(t, err)
-	require.Equal(t, int64(1), nextID)
-
-	upkeep = newUpkeep(registry, 3)
-	err = orm.UpsertUpkeep(&upkeep)
-	require.NoError(t, err)
-
-	nextID, err = orm.LowestUnsyncedID(registry.ID)
-	require.NoError(t, err)
-	require.Equal(t, int64(4), nextID)
-}
-
-func TestKeeperDB_SetLastRunInfoForUpkeepOnJob(t *testing.T) {
-	t.Parallel()
-	db, config, orm := setupKeeperDB(t)
-	ethKeyStore := cltest.NewKeyStore(t, db, config).Eth()
-
-	registry, j := cltest.MustInsertKeeperRegistry(t, db, orm, ethKeyStore, 0, 2, 20)
+	registry, j := cltest.MustInsertKeeperRegistry(t, db, orm, ethKeyStore, 0, 1, 20)
 	upkeep := cltest.MustInsertUpkeepForRegistry(t, db, config, registry)
 
-	// check normal behavior
-	err := orm.SetLastRunInfoForUpkeepOnJob(j.ID, upkeep.UpkeepID, 100, registry.FromAddress)
-	require.NoError(t, err)
+	// update
+	require.NoError(t, orm.SetLastRunInfoForUpkeepOnJob(j.ID, upkeep.UpkeepID, 100, registry.FromAddress))
 	assertLastRunHeight(t, db, upkeep, 100, 0)
-	// check that if we put in an unknown from address nothing breaks
-	err = orm.SetLastRunInfoForUpkeepOnJob(j.ID, upkeep.UpkeepID, 0, cltest.NewEIP55Address())
-	require.NoError(t, err)
+	// update to lower block not allowed
+	require.NoError(t, orm.SetLastRunInfoForUpkeepOnJob(j.ID, upkeep.UpkeepID, 0, registry.FromAddress))
 	assertLastRunHeight(t, db, upkeep, 100, 0)
+	// update to higher block allowed
+	require.NoError(t, orm.SetLastRunInfoForUpkeepOnJob(j.ID, upkeep.UpkeepID, 101, registry.FromAddress))
+	assertLastRunHeight(t, db, upkeep, 101, 0)
 }
